fix: Resolve compilation errors and complete test organization
- Fix TestData class: remove static modifier, add package and imports
- Add missing static imports for JUnit assertions in integration test
- Move testPersistReqRes_WhenFalse_DoesNotStoreData into Store Operations nested class
- Create Contains Operations nested class for contains-related tests
- All tests now properly organized with descriptive display names
- Verify all 23 tests (9 unit + 14 integration) pass successfully

Test structure now includes:
- Store Operations: 5 tests with clear display names
- Contains Operations: 2 tests for key existence checks
- Individual tests for other repository operations
